pyproject.toml
[tool.poetry]
authors = ["prasannaboga"]
description = ""
name = "project-cvb"
packages = [{include = "project_cvb"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "0.111.0"
google-generativeai = "0.7.1"
ipykernel = "6.29.4"
ipywidgets = "8.1.2"
openai = "1.35.5"
pillow = "10.3.0"
pydantic = "2.7.0"
pydantic-settings = "2.3.4"
python = "3.12.2"
python-dotenv = "1.0.1"
uvicorn = "0.29.0"

[tool.poetry.scripts]
hello_world = "project_cvb.scripts.hello_world:main"

[tool.custom_settings]
new_value = "new_value"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
